IEEE1516E data structs updated: - Fixed several encoding\decoding bugs related to the vc10 compiler. - Fixed getting a wrong pointer to the data in some structures. - Replaced 'long' by 'long long' for processing 64 bit values. Long is 4-bytes length on MS platforms.

+++ codebase/src/cpp/ieee1516e/src/types/encoding/BitHelpers.h
@@ -117,15 +117,15 @@ class BitHelpers
 		////////////////////////////////////////////////////////////
 		// Long Helpers
 		////////////////////////////////////////////////////////////
-		static void encodeLongBE( long value, char* buffer, size_t offset );
-		static long decodeLongBE( const char* buffer, size_t offset );
-		static void encodeLongBE( long value, std::vector<Octet>& buffer );
-		static long decodeLongBE( const std::vector<Octet>& buffer, size_t index )
+		static void encodeLongBE( long long value, char* buffer, size_t offset );
+		static long long decodeLongBE( const char* buffer, size_t offset );
+		static void encodeLongBE(long long value, std::vector<Octet>& buffer );
+		static long long decodeLongBE( const std::vector<Octet>& buffer, size_t index )
 			throw (EncoderException);
-		static void encodeLongLE( long value, char* buffer, size_t offset );
-		static long decodeLongLE( const char* buffer, size_t offset );
-		static void encodeLongLE( long value, std::vector<Octet>& buffer );
-		static long decodeLongLE( const std::vector<Octet>& buffer, size_t index )
+		static void encodeLongLE(long long value, char* buffer, size_t offset );
+		static long long decodeLongLE( const char* buffer, size_t offset );
+		static void encodeLongLE(long long value, std::vector<Octet>& buffer );
+		static long long decodeLongLE( const std::vector<Octet>& buffer, size_t index )
 			throw (EncoderException);
 
 		////////////////////////////////////////////////////////////
